Added upload_file_to_storage to azure_connector
Copied and reworked pyktok code to get the tiktok
json, which allows keeping track of the cookies.
The cookies were required to have permission to get
the video content from the download_url.

Without the need for pyktok to keep track of cookies,
pyktok may soon be replaced by custom code.